Compile and link under Debian GNU/Linux.
This should work for other Linux variants as well. The
ARM Ltd GCC port "arm-none-eabi-gcc" is required.  That
port targets Ubuntu via a PPA, but is also available in
Debian Jessie and Fedora 23.

As well as the Makefile the .gitignore is extended to
cover common Unix software development tools. If there
are additions then consider sending a patch to gitignore.io,
as that site generated most of the file.