Configuration for OpenOCD
OpenOCD is a open-source on-chip debugger implementation, see
<http://openocd.org/>. Version 0.9.0 or later is needed for
the ZodiacFX board.

This file works unaltered with the Atmel ICE re-sold by Northbound
Networks. Unplug the ICE's USB port and board's USB port, removing power
from both units. Plug the provided cable into the "SAM" port on the ICE
and into the JTAG headers on the ZodiacFX (noting that the marked wire
on the ribbon cable goes furthest away from the "JTAG" silk screened
label on the printed circuit board). Insert the USB port of the
ICE. Insert the USB port of the ZodiacFX.

From the same directory as this file, run "openocd". From another window
type "telnet localhost 4444" and the OpenOCD command line will
appear. Most commands are straight-forward, but the flash programming
commands are not: for convenience the flash commands are given in
comments in the openocd.conf file.

To use GDB say
  arm-none-eabi-gdb --eval-command="target remote localhost:3333"

If you are using OpenOCD from an internet-connected machine then
limit access to ports 3333, 4444, 6666. These port numbers can be
changed, see <http://openocd.org/doc/html/Daemon-Configuration.html>.