openflow: Cleanup of_receive() size/len usage
'size' is really the offset into 'packetbuffer', while
'len' is the amount of data in 'packetbuffer'. 'plen'
is the OF packet length.

Ideally OF message processing routines should only care
about the data for the specific message they are processing,
and do not need to know the offset within 'packetbuffer'
or the amount of data in 'packetbuffer'.